+++ app.py
@@ -5,7 +5,7 @@ from flask import Flask
 
 app_logger = logging.getLogger('yfinance-api')
 
-app_logger.setLevel(logging.DEBUG)
+app_logger.setLevel(logging.INFO)
 handler = logging.StreamHandler()
 formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 handler.setFormatter(formatter)